feat: moved primitives templates to BaseTemplate abstraction
