+++ dspy/backends/__init__.py
@@ -1,1 +1,3 @@
 from .template import TemplateBackend
+from .json import JSONBackend
+from .lm import LiteLM
