+++ dspy/backends/lm/litellm.py
@@ -5,11 +5,18 @@ from pydantic import Field
 
 from .base import BaseLM
 
+# Currently, LiteLLM logs way to much to the console in Info
+# I am setting this to WARNING to silence this
+# at some point, we will need to manage this appropriately with the new dspy logger
+import logging
+
+logger = logging.getLogger("LiteLLM")
+logger.setLevel(logging.WARNING)
 
 class LiteLM(BaseLM):
     STANDARD_PARAMS: dict[str, t.Any] = {
         "temperature": 0,
-        "max_tokens": 150,
+        "max_tokens": 500,
         "top_p": 1,
         "frequency_penalty": 0,
         "presence_penalty": 0,
@@ -34,6 +41,10 @@ class LiteLM(BaseLM):
 
         if type(response) != ModelResponse:
             raise AssertionError("Response from completion incorrect type/format")
+
+        # TODO: Move this to proper logging
+        if len([c for c in response.choices if c["finish_reason"] == "length"]) > 0:
+            print("Some of the generations are being limited by 'max_tokens', you may want to raise this value.")
 
         return [c["message"]["content"] for c in response.choices if c["finish_reason"] != "length"]
