+++ dspy/backends/templates/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseTemplate
+from .json import JSONTemplate
+from .text import TextTemplate
