+++ dspy/primitives/__init__.py
@@ -3,4 +3,3 @@ from .example import *
 from .prediction import *
 from .program import *
 from .python_interpreter import *
-from .template import *
